refactor(core): Cast size_t to long for Geom_Number arithmetic
Add static_cast<long> to size_t variables before division by Geom_Number.
Update CMakeLists.txt to add a new test target.
This avoids potential implicit conversion issues and improves type safety.

+++ Tests/geom_algorithms_test_edgecases_sweepline_minkowski_kdtree.cc
@@ -2263,8 +2263,8 @@ TEST_F(GeomAlgorithmsTest, SweepLineManySegmentsAtOnePoint)
   const size_t N = 10;
   for (size_t i = 0; i < N; ++i)
     {
-      Geom_Number angle = Geom_Number(2) * Geom_Number(M_PI) * Geom_Number(i)
-                          / Geom_Number(N);
+      Geom_Number angle = Geom_Number(2) * Geom_Number(M_PI) * Geom_Number(static_cast<long>(i))
+                          / Geom_Number(static_cast<long>(N));
       Point far(Geom_Number(5) + Geom_Number(10) * Geom_Number(cos(angle.get_d())),
                 Geom_Number(5) + Geom_Number(10) * Geom_Number(sin(angle.get_d())));
       segs.append(Segment(Point(5, 5), far));
